# [Host.AzureServiceBus] Create subscriptions in parallel