@EtherZa 2025-02-13
A quick 3.0.2 would be great tbh. It would vastly reduce my test feedback cycle.