@zarusz 2025-02-16
It has been released under 3.0.2.