@EtherZa 2025-02-17
> It has been released under 3.0.2.

Thank you @zarusz. 

Something is up with nuget. I can view and pull the packages from within both Visual Studio and Rider, but am unable to navigate to them on [nuget.org](https://www.nuget.org/packages/SlimMessageBus). Navigating to [3.0.1](https://www.nuget.org/packages/SlimMessageBus/3.0.1) or [3.0.2](https://www.nuget.org/packages/SlimMessageBus/3.0.2) directly redirects to 3.0.0.

There doesn't appear to be anything glaringly obvious in the nuspec, but perhaps another change is impacting it? Etiher way, I thought that you may like to know.

![image](https://github.com/user-attachments/assets/a28f1c44-ed19-43bb-a289-48d1122c12a5)